Initial .travis.yml addition
This adds .travis.yml and helper scripts for potential use on the
openstack-helm repo.  The workflow is as follows: install Glide as
a requirement for kubernetes-helm, build helm from source (this is
required to run tiller locally), deploy the kubeadm-ci container,
execute the config script, then run tiller locally. This allows us
to run `helm install --dry-run --debug {x}`, and also gives us the
ability to point to an external kubernetes cluster for actual
installs.